requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# project requirements 
numpy
pandas<2.0.0
torch
torchsummary
shap
scikit-learn
mpi4py
matplotlib
seaborn
tqdm
biopython
scipy
pylint
keras
tensorflow
tensorboard
shap
ipykernel
statannotations
audtorch
